pyproject.toml
[tool.poetry]
name = "tictactoe"
version = "0.1.0"
description = ""
authors = ["Eugene Morozov <jmv@emorozov.net>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
django-extensions = "^3.1.3"
drf-spectacular = "^0.18.2"
uvicorn = "^0.15.0"

[tool.poetry.dev-dependencies]
ipython = "^7.27.0"
flake8 = "^3.9.2"
pytest = "^6.2.5"
pytest-django = "^4.4.0"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
